+++ pinot-integration-tests/src/test/java/com/linkedin/pinot/integration/tests/HybridClusterIntegrationTest.java
@@ -83,7 +83,6 @@ public class HybridClusterIntegrationTest extends BaseClusterIntegrationTest {
   private String tableName;
 
   private KafkaServerStartable kafkaStarter;
-  private HelixManager _zkHelixManager;
 
   protected void setSegmentCount(int segmentCount) {
     this.segmentCount = segmentCount;
@@ -102,6 +101,7 @@ public class HybridClusterIntegrationTest extends BaseClusterIntegrationTest {
   }
 
   // In case inherited tests set up a different table name they can override this method.
+  @Override
   protected String getTableName() {
     return tableName;
   }
@@ -338,123 +338,10 @@ public class HybridClusterIntegrationTest extends BaseClusterIntegrationTest {
     }
   }
 
-  protected void ensureZkHelixManagerIsInitialized() {
-    if (_zkHelixManager == null) {
-      _zkHelixManager = HelixManagerFactory.getZKHelixManager(getHelixClusterName(), "test_instance", InstanceType.SPECTATOR,
-          ZkStarter.DEFAULT_ZK_STR);
-      try {
-        _zkHelixManager.connect();
-      } catch (Exception e) {
-        throw new RuntimeException(e);
-      }
-    }
-  }
-
+  @Override
   @Test
   public void testInstanceShutdown() {
-    ensureZkHelixManagerIsInitialized();
-
-    HelixAdmin clusterManagmentTool = _zkHelixManager.getClusterManagmentTool();
-    String clusterName = _zkHelixManager.getClusterName();
-    List<String> instances = clusterManagmentTool.getInstancesInCluster(clusterName);
-    Assert.assertFalse(instances.isEmpty(), "List of instances should not be empty");
-
-    // Mark all instances in the cluster as shutting down
-    for (String instance : instances) {
-      InstanceConfig instanceConfig = clusterManagmentTool.getInstanceConfig(clusterName, instance);
-      instanceConfig.getRecord().setBooleanField(CommonConstants.Helix.IS_SHUTDOWN_IN_PROGRESS, true);
-      clusterManagmentTool.setInstanceConfig(clusterName, instance, instanceConfig);
-    }
-
-    // Check that the routing table is empty
-    checkForEmptyRoutingTable(true, "Routing table is not empty after marking all instances as shutting down");
-
-    // Mark all instances as not shutting down
-    for (String instance : instances) {
-      InstanceConfig instanceConfig = clusterManagmentTool.getInstanceConfig(clusterName, instance);
-      instanceConfig.getRecord().setBooleanField(CommonConstants.Helix.IS_SHUTDOWN_IN_PROGRESS, false);
-      clusterManagmentTool.setInstanceConfig(clusterName, instance, instanceConfig);
-    }
-
-    // Check that the routing table is not empty
-    checkForEmptyRoutingTable(false, "Routing table is empty after marking all instances as not shutting down");
-  }
-
-  protected void checkForEmptyRoutingTable(final boolean checkForEmpty, String message) {
-    waitForRoutingTablePredicate(new Function<JSONArray, Boolean>() {
-      @Override
-      public Boolean apply(JSONArray input) {
-        try {
-          int tableCount = input.length();
-
-          // Routing table should have an entry for this table
-          if (tableCount == 0) {
-            return false;
-          }
-
-          // Each routing table entry for this table should have not have any server to segment mapping
-          for (int i = 0; i < tableCount; i++) {
-            JSONObject tableRouting = input.getJSONObject(i);
-            String tableName = tableRouting.getString("tableName");
-            if (tableName.startsWith(getTableName())) {
-              JSONArray routingTableEntries = tableRouting.getJSONArray("routingTableEntries");
-              int routingTableEntryCount = routingTableEntries.length();
-              for (int j = 0; j < routingTableEntryCount; j++) {
-                JSONObject routingTableEntry = routingTableEntries.getJSONObject(j);
-
-                boolean hasServerToSegmentMappings = routingTableEntry.keys().hasNext();
-
-                if (hasServerToSegmentMappings && checkForEmpty) {
-                  return false;
-                }
-
-                if (!hasServerToSegmentMappings && !checkForEmpty) {
-                  return false;
-                }
-              }
-            }
-          }
-
-          return true;
-        } catch (JSONException e) {
-          LOGGER.warn("Caught exception while reading the routing table, will retry", e);
-          return false;
-        }
-      }
-    }, 15000, message);
-
-  }
-
-  /**
-   * Wait for a predicate on the routing table to become true.
-   * @param predicate true when the routing table condition is met
-   * @param timeout Timeout for the predicate to become true
-   * @param message Message to display if the predicate does not become true after the timeout expires
-   */
-  protected void waitForRoutingTablePredicate(Function<JSONArray, Boolean> predicate, long timeout, String message) {
-    long endTime = System.currentTimeMillis() + timeout;
-    boolean isPredicateMet = false;
-    JSONObject routingTableSnapshot = null;
-    while (System.currentTimeMillis() < endTime && !isPredicateMet) {
-      try {
-        routingTableSnapshot = getDebugInfo("debug/routingTable/" + getTableName());
-
-        if (routingTableSnapshot != null) {
-          JSONArray routingTableSnapshotJson = routingTableSnapshot.getJSONArray("routingTableSnapshot");
-          if (routingTableSnapshotJson != null) {
-            isPredicateMet = predicate.apply(routingTableSnapshotJson);
-          }
-        } else {
-          LOGGER.warn("Got null routing table snapshot, retrying");
-        }
-      } catch (Exception e) {
-        // Will retry in a bit
-      }
-
-      Uninterruptibles.sleepUninterruptibly(500, TimeUnit.MILLISECONDS);
-    }
-
-    Assert.assertTrue(isPredicateMet, message + ", last routing table snapshot is " + routingTableSnapshot.toString());
+    super.testInstanceShutdown();
   }
 
   @AfterClass
