+++ pinot-integration-tests/src/test/java/com/linkedin/pinot/integration/tests/UploadRefreshDeleteIntegrationTest.java
@@ -372,6 +372,11 @@ public class UploadRefreshDeleteIntegrationTest extends BaseClusterIntegrationTe
     // Ignored.
   }
 
+  @Override
+  protected String getTableName() {
+    return tableName;
+  }
+
   @AfterClass
   public void tearDown() {
     stopServer();
